DDLS Exisiting Model Added

+++ src/Modules/Execution/Execution.Application/Features/DailyLabor/Commands/DailyLaborCommands.cs
@@ -4,5 +4,5 @@ using Himapp.Execution.Application.Features.DailyLabor.Models;
 namespace Himapp.Execution.Application.Features.DailyLabor.Commands;
 
 public sealed record CreateDailyLaborCommand(CreateDailyLaborRequest Request) : IRequest<DailyLaborModel>;
-public sealed record UpdateDailyLaborCommand(long Id, UpdateDailyLaborRequest Request) : IRequest<DailyLaborModel?>;
-public sealed record DeleteDailyLaborCommand(long Id) : IRequest<bool>;
+public sealed record UpdateDailyLaborCommand(int Id, UpdateDailyLaborRequest Request) : IRequest<DailyLaborModel?>;
+public sealed record DeleteDailyLaborCommand(int Id) : IRequest<bool>;
